+++ logic/src/appleMain/kotlin/com/wire/kalium/logic/di/PlatformUserStorageProvider.kt
@@ -27,7 +27,7 @@ import com.wire.kalium.persistence.kmmSettings.UserPrefBuilder
 import com.wire.kalium.util.KaliumDispatcherImpl
 
 internal actual class PlatformUserStorageProvider actual constructor() : UserStorageProvider() {
-    override fun create(userId: UserId, shouldEncryptData: Boolean, platformProperties: PlatformUserStorageProperties): UserStorage {
+    actual override fun create(userId: UserId, shouldEncryptData: Boolean, platformProperties: PlatformUserStorageProperties): UserStorage {
         val userIdEntity = userId.toDao()
         val pref = UserPrefBuilder(userIdEntity, platformProperties.rootPath, shouldEncryptData)
         val database = userDatabaseBuilder(
