+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/data/asset/KaliumFileSystem.kt
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2024 Wire Swiss GmbH
+ * Copyright (C) 2025 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,34 +15,23 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
-
 package com.wire.kalium.logic.data.asset
 
-import com.wire.kalium.util.KaliumDispatcher
-import kotlinx.coroutines.withContext
-import okio.FileSystem.Companion.SYSTEM
 import okio.Path
-import okio.Path.Companion.toPath
 import okio.Sink
 import okio.Source
-import okio.buffer
-import okio.use
 
 @Suppress("TooManyFunctions")
-actual class KaliumFileSystemImpl actual constructor(
-    private val dataStoragePaths: DataStoragePaths,
-    private val dispatcher: KaliumDispatcher
-) : KaliumFileSystem {
-
+interface KaliumFileSystem {
     /**
      * Provides the root of the cache path, used to store temporary files
      */
-    actual override val rootCachePath: Path = dataStoragePaths.cachePath.value.toPath()
+    val rootCachePath: Path
 
     /**
      * Provides the root of the current user database path, used to store all the Database information.
      */
-    actual override val rootDBPath: Path = dataStoragePaths.dbPath.value.toPath()
+    val rootDBPath: Path
 
     /**
      * Opens an output stream that will be used to write the data on the given [outputPath]
@@ -50,77 +39,75 @@ actual class KaliumFileSystemImpl actual constructor(
      * @param mustCreate whether to force the creation of the outputPath if it doesn't exist on the current file system
      * @return the [Sink] stream of data to be written
      */
-    actual override fun sink(outputPath: Path, mustCreate: Boolean): Sink = SYSTEM.sink(outputPath, mustCreate)
+    fun sink(outputPath: Path, mustCreate: Boolean = false): Sink
 
     /**
      * Creates an input stream that will be used to read the data from the given [inputPath]
      * @param inputPath the path from where the data will be read
      * @return the [Source] stream of data to be read
      */
-    actual override fun source(inputPath: Path): Source = SYSTEM.source(inputPath)
+    fun source(inputPath: Path): Source
 
     /**
      * It will create the provided [dir] in the current file system along with the needed subdirectories if they were not created previously
      */
-    actual override fun createDirectories(dir: Path) = SYSTEM.createDirectories(dir)
+    fun createDirectories(dir: Path)
 
     /**
      * It will make sure the given [dir] gets created on the file system
      * @param mustCreate whether it is certain that [dir] doesn't exist and will need to be created
      */
-    actual override fun createDirectory(dir: Path, mustCreate: Boolean) = SYSTEM.createDirectory(dir, mustCreate)
+    fun createDirectory(dir: Path, mustCreate: Boolean = false)
 
     /**
-     * This will delete the content of the given [path]
+     * This will delete the content of the given file [path]
      * @param path the path to be deleted
      * @param mustExist whether it is certain that [path] exists before the deletion
      */
-    actual override fun delete(path: Path, mustExist: Boolean) = SYSTEM.delete(path, mustExist)
+    fun delete(path: Path, mustExist: Boolean = false)
 
     /**
      * This will delete recursively the given [dir] and all its content
      * @param dir the directory to be deleted
      * @param mustExist whether it is certain that [dir] exists before the deletion
      */
-    actual override fun deleteContents(dir: Path, mustExist: Boolean) = SYSTEM.deleteRecursively(dir, mustExist)
+    fun deleteContents(dir: Path, mustExist: Boolean = false)
 
     /**
      * Checks whether the given [path] is already created and exists on the current file system
      * @return whether the given [path] exists in the current file system
      */
-    actual override fun exists(path: Path): Boolean = SYSTEM.exists(path)
+    fun exists(path: Path): Boolean
 
     /**
      * Copies effectively the content of [sourcePath] into [targetPath]
      * @param sourcePath the path of the content to be copied
      * @param targetPath the destination path where the data will be copied into
      */
-    actual override fun copy(sourcePath: Path, targetPath: Path) = SYSTEM.copy(sourcePath, targetPath)
+    fun copy(sourcePath: Path, targetPath: Path)
 
     /**
-     * Creates a temporary path if it didn't exist before and returns it if successful
+     * Creates a temporary path if it didn't exist before and returns it if successful.
      * @param pathString a predefined temp path string. If not provided the temporary folder will be created with a default path
      */
-    actual override fun tempFilePath(pathString: String?): Path {
-        val filePath = pathString ?: "temp_file_path"
-        return "$rootCachePath/$filePath".toPath()
-    }
+    fun tempFilePath(pathString: String? = null): Path
 
     /**
      * Creates a persistent path on the internal storage folder of the file system if it didn't exist before and returns it if successful
      * @param assetName the asset path string
      */
-    actual override fun providePersistentAssetPath(assetName: String): Path = "${dataStoragePaths.assetStoragePath.value}/$assetName".toPath()
+    fun providePersistentAssetPath(assetName: String): Path
+
+    /**
+     * Fetches the persistent [Path] of the current user's avatar in the [KaliumFileSystem]
+     */
+    fun selfUserAvatarPath(): Path
 
     /**
      * Reads the data of the given path as a byte array
      * @param inputPath the path pointing to the stored data
      */
-    actual override suspend fun readByteArray(inputPath: Path): ByteArray = withContext(dispatcher.io) {
-        source(inputPath).buffer().use { bufferedFileSource ->
-            bufferedFileSource.readByteArray()
-        }
-    }
+    suspend fun readByteArray(inputPath: Path): ByteArray
 
     /**
      * Writes the data contained on [dataSource] into the provided [outputSink]
@@ -128,27 +115,12 @@ actual class KaliumFileSystemImpl actual constructor(
      * @param dataSource the data source that kaliumFileSystem will read from to write the data to the [outputSink]
      * @return the number of bytes written
      */
-    actual override suspend fun writeData(outputSink: Sink, dataSource: Source): Long {
-        var byteCount = 0L
-        withContext(dispatcher.io) {
-            outputSink.use { sink ->
-                sink.buffer().use { bufferedFileSink ->
-                    byteCount = bufferedFileSink.writeAll(dataSource)
-                }
-            }
-        }
-        return byteCount
-    }
-
-    /**
-     * Fetches the persistent [Path] of the current user's avatar in the [KaliumFileSystem]
-     */
-    actual override fun selfUserAvatarPath(): Path = providePersistentAssetPath("self_user_avatar.jpg")
+    suspend fun writeData(outputSink: Sink, dataSource: Source): Long
 
     /**
      * Provides a list of paths found in the given [dir] path from where the call is being invoked.
      * @param dir the path from where the list of paths will be fetched
      * @return the list of paths found.
      */
-    actual override suspend fun listDirectories(dir: Path): List<Path> = SYSTEM.list(dir)
+    suspend fun listDirectories(dir: Path): List<Path>
 }
