+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/di/PlatformUserStorageProvider.kt
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2024 Wire Swiss GmbH
+ * Copyright (C) 2025 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,13 +15,14 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
-
 package com.wire.kalium.logic.di
 
 import com.wire.kalium.logic.data.user.UserId
 
-abstract class RootPathsProvider(val rootPath: String) {
-    abstract fun rootAccountPath(userId: UserId): String
-    abstract fun rootProteusPath(userId: UserId): String
-    abstract fun rootMLSPath(userId: UserId): String
+internal expect class PlatformUserStorageProvider constructor() : UserStorageProvider {
+    override fun create(
+        userId: UserId,
+        shouldEncryptData: Boolean,
+        platformProperties: PlatformUserStorageProperties
+    ): UserStorage
 }
