+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/di/UserStorageProvider.kt
@@ -43,5 +43,4 @@ abstract class UserStorageProvider {
     fun clearInMemoryUserStorage(userId: UserId) = inMemoryUserStorage.remove(userId)
 }
 
-internal expect class PlatformUserStorageProvider constructor() : UserStorageProvider
 expect class PlatformUserStorageProperties
