+++ logic/src/jvmMain/kotlin/com/wire/kalium/logic/di/PlatformRootPathsProvider.kt
@@ -21,8 +21,8 @@ package com.wire.kalium.logic.di
 import com.wire.kalium.logic.data.user.UserId
 
 actual class PlatformRootPathsProvider actual constructor(rootPath: String) : RootPathsProvider(rootPath) {
-    override fun rootAccountPath(userId: UserId): String = "$rootPath/${userId.domain}/${userId.value}"
-    override fun rootProteusPath(userId: UserId): String = "${rootAccountPath(userId)}/proteus"
-    override fun rootMLSPath(userId: UserId): String = "${rootAccountPath(userId)}/mls"
+    actual override fun rootAccountPath(userId: UserId): String = "$rootPath/${userId.domain}/${userId.value}"
+    actual override fun rootProteusPath(userId: UserId): String = "${rootAccountPath(userId)}/proteus"
+    actual override fun rootMLSPath(userId: UserId): String = "${rootAccountPath(userId)}/mls"
 
 }
