+++ logic/src/jvmMain/kotlin/com/wire/kalium/logic/di/PlatformUserStorageProvider.kt
@@ -28,7 +28,7 @@ import com.wire.kalium.persistence.kmmSettings.UserPrefBuilder
 import com.wire.kalium.util.KaliumDispatcherImpl
 
 internal actual class PlatformUserStorageProvider : UserStorageProvider() {
-    override fun create(
+    actual override fun create(
         userId: UserId,
         shouldEncryptData: Boolean,
         platformProperties: PlatformUserStorageProperties
